requirements.txt
#
# This file is autogenerated by pip-compile
# To update, run:
#
#    pip-compile
#
amqp==2.6.1               # via kombu
appdirs==1.4.4            # via black
asgiref==3.3.0            # via django
attrs==20.2.0             # via black
autoflake==1.4            # via -r requirements.in
billiard==3.6.3.0         # via celery
black==19.10b0            # via -r requirements.in
celery==4.4.7             # via -r requirements.in
certifi==2020.6.20        # via requests
chardet==3.0.4            # via requests
click==7.1.2              # via black
crontab==0.22.9           # via -r requirements.in
dj_database_url==0.5.0    # via -r requirements.in
django-crispy-forms==1.9.2  # via -r requirements.in
django-filter==2.4.0      # via -r requirements.in
django==3.0.4             # via -r requirements.in, django-filter, djangorestframework
djangorestframework==3.11.0  # via -r requirements.in
envparse==0.2.0           # via -r requirements.in
gunicorn==19.9.0          # via -r requirements.in
idna==2.10                # via requests
invoke==1.4.1             # via -r requirements.in
isort==4.3.21             # via -r requirements.in
kombu==4.6.11             # via celery
pathspec==0.8.0           # via black
psycopg2-binary==2.8.4    # via -r requirements.in
pyflakes==2.2.0           # via autoflake
pytz==2020.1              # via celery, django
regex==2020.10.15         # via black
requests==2.24.0          # via -r requirements.in
sqlparse==0.4.1           # via django
toml==0.10.1              # via black
typed-ast==1.4.1          # via black
urllib3==1.25.11          # via requests
vine==1.3.0               # via amqp, celery
